+++ lib/model/history_entry/repository.dart
@@ -1,20 +1,20 @@
 import 'dart:async';
 
-import 'package:beerstory/model/history_entry/database.dart';
+import 'package:beerstory/model/database.dart';
 import 'package:beerstory/model/history_entry/history_entry.dart';
 import 'package:beerstory/model/repository.dart';
-import 'package:flutter/material.dart';
+import 'package:drift/drift.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 
 /// The history provider.
-final historyProvider = AsyncNotifierProvider<History, List<HistoryEntry>>(
-  History.new,
-);
+final historyProvider = AsyncNotifierProvider<History, List<HistoryEntry>>(History.new);
 
 /// The repository that handles history entries.
-class History extends Repository<HistoryEntry> {
-  Future<void> aborb(HistoryEntry entry) async {
-    List<HistoryEntry> entries = await ref.read(databaseProvider).list(date: entry.date);
+class History extends Repository<HistoryEntry> with DatabaseRepository<HistoryEntry, DriftHistoryEntry, HistoryEntries> {
+  /// Finds the entry with the given [entry.beerUuid] and calls [absorbEntry] on it.
+  /// Only calls [add] if no entry is found.
+  Future<void> absorb(HistoryEntry entry) async {
+    List<HistoryEntry> entries = await future;
     for (HistoryEntry historyEntry in entries) {
       if (historyEntry.beerUuid == entry.beerUuid) {
         await change(historyEntry.absorbEntry(entry));
@@ -31,8 +31,33 @@ class History extends Repository<HistoryEntry> {
   }
 
   @override
-  @protected
-  AutoDisposeProvider<HistoryEntriesDatabase> get databaseProvider => historyEntriesDatabaseProvider;
+  Future<void> change(HistoryEntry object) async {
+    await super.change(object);
+    ref.read(lastInsertedBeerProvider.notifier).state = object.beerUuid;
+  }
+
+  @override
+  TableInfo<HistoryEntries, DriftHistoryEntry> getTable(Database database) => database.historyEntries;
+
+  @override
+  Insertable<DriftHistoryEntry> toInsertable(HistoryEntry object) => DriftHistoryEntry(
+    uuid: object.uuid,
+    date: object.date,
+    beerUuid: object.beerUuid,
+    quantity: object.quantity,
+    times: object.times,
+    moreThanQuantity: object.moreThanQuantity,
+  );
+
+  @override
+  HistoryEntry toObject(DriftHistoryEntry insertable) => HistoryEntry(
+    uuid: insertable.uuid,
+    date: insertable.date,
+    beerUuid: insertable.beerUuid,
+    quantity: insertable.quantity,
+    times: insertable.times,
+    moreThanQuantity: insertable.moreThanQuantity,
+  );
 }
 
 /// The last inserted beer provider.
